pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-binance>=1.0.29",
]
